+++ tests/test_utils_user_management.py
@@ -96,7 +96,7 @@ def test_user_management_helper_functions(tmp_app):  # NOQA
     ]
     assert list_users() == expected_content
 
-    # Test deleting a user
+    # Test deleting users.
     from dtool_lookup_server.utils import delete_users
 
     users_to_delete = [
@@ -125,6 +125,32 @@ def test_user_management_helper_functions(tmp_app):  # NOQA
         {
             "username": data_champion_username,
             "is_admin": False,
+            "search_permissions_on_base_uris": [],
+            "register_permissions_on_base_uris": []
+        },
+    ]
+    assert list_users() == expected_content
+
+    # Test updating users admin privileges.
+    from dtool_lookup_server.utils import update_users
+
+    users_to_update = [
+        {"username": admin_username},  # The is_admin value defaults to False.
+        {"username": data_champion_username, "is_admin": True},
+        {"username": standard_user_username},  # Not in system so ignored.
+    ]
+    update_users(users_to_update)
+
+    expected_content = [
+        {
+            "username": admin_username,
+            "is_admin": False,
+            "search_permissions_on_base_uris": [],
+            "register_permissions_on_base_uris": []
+        },
+        {
+            "username": data_champion_username,
+            "is_admin": True,
             "search_permissions_on_base_uris": [],
             "register_permissions_on_base_uris": []
         },